+++ src/main/java/com/elmakers/mine/bukkit/wand/WandUpgradePath.java
@@ -3,6 +3,10 @@ package com.elmakers.mine.bukkit.wand;
 import org.apache.commons.lang.StringUtils;
 import org.bukkit.configuration.ConfigurationSection;
 
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 import java.util.TreeMap;
 
 /**
@@ -12,8 +16,50 @@ import java.util.TreeMap;
  * Upgrading is generally done by spending XP on an enchanting table.
  */
 public class WandUpgradePath {
+    private static Map<String, WandUpgradePath> paths = new HashMap<String, WandUpgradePath>();
+
     private TreeMap<Integer, WandLevel> levelMap = null;
     private int[] levels = null;
+    private final String key;
+
+    public WandUpgradePath(String key, ConfigurationSection template) {
+        this.key = key;
+
+        // Parse defined levels
+        levelMap = new TreeMap<Integer, WandLevel>();
+        String[] levelStrings = StringUtils.split(template.getString("levels"), ",");
+        levels = new int[levelStrings.length];
+        for (int i = 0; i < levels.length; i++) {
+            levels[i] = Integer.parseInt(levelStrings[i]);
+        }
+
+        for (int level = 1; level <= levels[levels.length - 1]; level++) {
+            // TODO: Could this be optimized?
+            int levelIndex;
+            int nextLevelIndex = 0;
+            float distance = 1;
+            for (levelIndex = 0; levelIndex < levels.length; levelIndex++) {
+                if (level == levels[levelIndex] || levelIndex == levels.length - 1) {
+                    nextLevelIndex = levelIndex;
+                    distance = 0;
+                    break;
+                }
+
+                if (level > levels[levelIndex]) {
+                    nextLevelIndex = levelIndex + 1;
+                    int previousLevel = levels[levelIndex];
+                    int nextLevel = levels[nextLevelIndex];
+                    distance = (float)(level - previousLevel) / (float)(nextLevel - previousLevel);
+                }
+            }
+
+            levelMap.put(level, new WandLevel(template, levelIndex, nextLevelIndex, distance));
+        }
+    }
+
+    public String getKey() {
+        return key;
+    }
 
     public WandLevel getLevel(int level) {
         if (levelMap == null) return null;
@@ -29,17 +75,39 @@ public class WandUpgradePath {
         return levelMap.get(level);
     }
 
-    public void mapLevels(ConfigurationSection template) {
-        // Parse defined levels
-        levelMap = new TreeMap<Integer, WandLevel>();
-        String[] levelStrings = StringUtils.split(template.getString("levels"), ",");
-        levels = new int[levelStrings.length];
-        for (int i = 0; i < levels.length; i++) {
-            levels[i] = Integer.parseInt(levelStrings[i]);
-        }
-
-        for (int level = 1; level <= levels[levels.length - 1]; level++) {
-            // levelMap.put(level, new WandLevel(level, template));
+    public static void loadPaths(ConfigurationSection configuration) {
+        paths.clear();
+        Set<String> pathKeys = configuration.getKeys(false);
+        for (String key : pathKeys)
+        {
+            WandUpgradePath path = new WandUpgradePath(key, configuration.getConfigurationSection(key));
+            paths.put(key, path);
         }
     }
+
+    public static Set<String> getPathKeys() {
+        return paths.keySet();
+    }
+
+    public static WandUpgradePath getPath(String key) {
+        return paths.get(key);
+    }
+
+    public int getMaxLevel() {
+        if (levels == null) return 0;
+
+        return Math.min(levels[levels.length - 1], WandLevel.maxLevel);
+    }
+
+    public Set<Integer> getLevels() {
+        if (levels == null) return null;
+        Set<Integer> filteredLevels = new HashSet<Integer>();
+        for (Integer level : levels) {
+            if (level >= WandLevel.minLevel && level <= WandLevel.maxLevel) {
+                filteredLevels.add(level);
+            }
+        }
+        return filteredLevels;
+    }
+
 }